requirements.txt
requests
asyncio
aiohttp
python-dotenv